// lib/_TC/_SeD/enums/BrdBlm_TC_SeD_eSlidingTrackFinish.ts
/** ---------------------------------------------------------------------------
 * @module [BrdBlm_TC_SeD_V_ST]
 * @author [APG] Angeli Paolo Giusto 
 * @version 0.1 APG 20231114
 * @version 0.2 APG 20231226 Estratto nel suo proprio file
 * ----------------------------------------------------------------------------
 */

/**
 * Tipologie della finitura delle guide per i sistemi di scorrimento dei portoni
 * sezionali a scorrimento verticale
 */
export enum BrdBlm_TC_SeD_eSlidingTrackFinish {

  GALVANIZED = "glv",
  WHITE = "wht",
  BLACK = "blk"

}

/*! ---------------------------------------------------------------------------
 * @copyright Breda Sistemi industriali S.p.A., 2023 - http://bredasys.com
 * All rights reserved
 * @licence You cannot host, display, distribute or share this Work in any
 * form, both physical and digital. You cannot use this Work in any commercial
 * or non-commercial product, website or project. You cannot sell this Work
 * and you cannot mint an NFTs out of it.
 * ---------------------------------------------------------------------------
 */